week_one_sdk/requirements.txt
google-genai>=1.18.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0